double precision on cuda device

+++ cuda/cuda_kernel.cu
@@ -3,33 +3,33 @@
 
 #include "../common/globals.h"
 
-__device__ float* g_data[2];
+__device__ double* g_data[2];
 __device__ int maximum_index;
 __device__ int data_size[2];
-__device__ float* g_weights[2];
-__device__ float lambda;
+__device__ double* g_weights[2];
+__device__ double lambda;
 __device__ int max_p_index;
-__device__ float max_p;
+__device__ double max_p;
 __device__ int max_q_index;
-__device__ float max_q;
-__device__ float dot_xi_yi;		 // <x_i, y_i >
-__device__ float dot_xi_xi;		 // <x_i, x_i >
-__device__ float dot_yi_yi;		 // <y_i, y_i >
-__device__ float* distance;
-__device__ float* rho;
-__device__ float* dot_xi_x;
-__device__ float* dot_yi_x;
-__device__ float* dot_xi_y;
-__device__ float* dot_yi_y;
-__device__ float* dot_same[2];
+__device__ double max_q;
+__device__ double dot_xi_yi;		 // <x_i, y_i >
+__device__ double dot_xi_xi;		 // <x_i, x_i >
+__device__ double dot_yi_yi;		 // <y_i, y_i >
+__device__ double* distance;
+__device__ double* rho;
+__device__ double* dot_xi_x;
+__device__ double* dot_yi_x;
+__device__ double* dot_xi_y;
+__device__ double* dot_yi_y;
+__device__ double* dot_same[2];
 
-__device__ float* get_element(int id, int set);
+__device__ double* get_element(int id, int set);
 
 __device__ struct svm_parameter param;
 
-__device__ float dot(float* px, float *py)
+__device__ double dot(double* px, double *py)
 {
-	float sum = 0.0;
+	double sum = 0.0;
 	int i;
 	for(i=0; i< maximum_index; i++)
 	{
@@ -39,9 +39,9 @@ __device__ float dot(float* px, float *py)
 }
 
 
-inline float powi(float base, int times)
+inline double powi(double base, int times)
 {
-	float tmp = base, ret = 1.0;
+	double tmp = base, ret = 1.0;
 
 	int t;
 	for(t=times; t>0; t/=2)
@@ -52,20 +52,20 @@ inline float powi(float base, int times)
 	return ret;
 }
 
-__device__ float kernel_linear(int set1, int element1, int set2, int element2) //todo: als template implementieren
+__device__ double kernel_linear(int set1, int element1, int set2, int element2) //todo: als template implementieren
 {
-	float* px = &(g_data[set1][ element1 * maximum_index ]);
-	float* py = &(g_data[set2][ element2 * maximum_index ]);
+	double* px = &(g_data[set1][ element1 * maximum_index ]);
+	double* py = &(g_data[set2][ element2 * maximum_index ]);
 
-	float ret = dot(px, py );
+	double ret = dot(px, py );
 	if(set1 == set2 && element1 == element2)
 		ret += 1.0/param.C;
 	return ret;
 }
 
-__device__ float power(float base, int exponent) { //todo: effizienter berechnen? (squaring, bitshifts)
+__device__ double power(double base, int exponent) { //todo: effizienter berechnen? (squaring, bitshifts)
 	int i;
-	float res = base; 
+	double res = base; 
 	for(i=0;i<exponent;i++) 
 	{
 		res  *= base;
@@ -73,27 +73,27 @@ __device__ float power(float base, int exponent) { //todo: effizienter berechnen
 	return res;
 }
 
-__device__ float kernel_poly(int set1, int element1, int set2, int element2)
+__device__ double kernel_poly(int set1, int element1, int set2, int element2)
 {
-	float* px = &(g_data[set1][ element1 * maximum_index ]);
-	float* py = &(g_data[set2][ element2 * maximum_index ]);
+	double* px = &(g_data[set1][ element1 * maximum_index ]);
+	double* py = &(g_data[set2][ element2 * maximum_index ]);
 
-	float ret = power(param.gamma*dot(px, py )+param.coef0,param.degree);
+	double ret = power(param.gamma*dot(px, py )+param.coef0,param.degree);
 	if(set1 == set2 && element1 == element2)
 		ret += 1.0/param.C;
 	return ret;
 }
 
-__device__ float kernel_rbf(int set1, int element1, int set2, int element2)
+__device__ double kernel_rbf(int set1, int element1, int set2, int element2)
 {
-	float* px = &(g_data[set1][ element1 * maximum_index ]);
-	float* py = &(g_data[set2][ element2 * maximum_index ]);
+	double* px = &(g_data[set1][ element1 * maximum_index ]);
+	double* py = &(g_data[set2][ element2 * maximum_index ]);
 
-	float dots = ( dot(px, px)+
+	double dots = ( dot(px, px)+
 						dot(py, py)-2*
 						dot(px, py)); //todo: dot(x,x) vorberechnen??
-	float wgamma = -param.gamma*dots;
-	float wexp = exp(wgamma);
+	double wgamma = -param.gamma*dots;
+	double wexp = exp(wgamma);
 
 	if(set1 == set2 && element1 == element2)
 		wexp += 1.0/param.C;
@@ -101,18 +101,18 @@ __device__ float kernel_rbf(int set1, int element1, int set2, int element2)
 
 }
 
-__device__ float kernel_sigmoid(int set1, int element1, int set2, int element2)
+__device__ double kernel_sigmoid(int set1, int element1, int set2, int element2)
 {
-	float* px = &(g_data[set1][ element1 * maximum_index ]);
-	float* py = &(g_data[set2][ element2 * maximum_index ]);
+	double* px = &(g_data[set1][ element1 * maximum_index ]);
+	double* py = &(g_data[set2][ element2 * maximum_index ]);
 
-	float ret = tanh(param.gamma*dot(px, py)+param.coef0);
+	double ret = tanh(param.gamma*dot(px, py)+param.coef0);
 	if(set1 == set2 && element1 == element2)
 		ret += 1.0/param.C;
 	return ret;
 }
 
-__device__ float kernel(int set1, int element1, int set2, int element2)
+__device__ double kernel(int set1, int element1, int set2, int element2)
 {
 	switch(param.kernel_type)
 	{
@@ -130,7 +130,7 @@ __device__ float kernel(int set1, int element1, int set2, int element2)
 	}
 }
 
-__device__ int find_max(int p, float *dot_yi_x, float* dot_xi_x, float dot_xi_yi, float dot_xi_xi, float *max_p)
+__device__ int find_max(int p, double *dot_yi_x, double* dot_xi_x, double dot_xi_yi, double dot_xi_xi, double *max_p)
 {
 	// find max
 	int max_p_index = -1;
@@ -139,7 +139,7 @@ __device__ int find_max(int p, float *dot_yi_x, float* dot_xi_x, float dot_xi_yi
 	int i;
 	for (i=0; i<data_size[p]; i++)
 	{
-		float sum = dot_yi_x[i] - dot_xi_x[i] - dot_xi_yi + dot_xi_xi;
+		double sum = dot_yi_x[i] - dot_xi_x[i] - dot_xi_yi + dot_xi_xi;
 		if(sum > *max_p)
 		{
 			*max_p = sum;
@@ -152,16 +152,16 @@ __device__ int find_max(int p, float *dot_yi_x, float* dot_xi_x, float dot_xi_yi
 
 template <unsigned int blockSize>
 								 //todo: set und first als template parameter?
-__global__ void reduce6(int *g_data_index, float *g_data_value, unsigned int set, int first, int data_size1)
+__global__ void reduce6(int *g_data_index, double *g_data_value, unsigned int set, int first, int data_size1)
 {
 	__shared__ int sdata_index[blockSize];
-	__shared__ float sdata_value[blockSize];
+	__shared__ double sdata_value[blockSize];
 	unsigned int tid = threadIdx.x;
 	unsigned int i = blockIdx.x*(blockSize*2) + tid;
 
 	if(first == 1)
 	{
-		float value1;
+		double value1;
 		if(set == 0)
 			value1 = dot_yi_x[i] - dot_xi_x[i] - dot_xi_yi + dot_xi_xi;
 		else
@@ -172,7 +172,7 @@ __global__ void reduce6(int *g_data_index, float *g_data_value, unsigned int set
 			if( i + blockSize < data_size[set])
 			{
 	
-				float value2;
+				double value2;
 				if(set == 0)
 					value2 = dot_yi_x[i+blockSize] - dot_xi_x[i+blockSize] - dot_xi_yi + dot_xi_xi;
 				else
@@ -201,8 +201,8 @@ __global__ void reduce6(int *g_data_index, float *g_data_value, unsigned int set
 	{
 		if( i + blockSize < data_size1)
 		{
-			float value1 = g_data_value[i];
-			float value2 = g_data_value[i+blockSize];
+			double value1 = g_data_value[i];
+			double value2 = g_data_value[i+blockSize];
 			if(value1>value2)
 			{
 				sdata_value[tid] = value1;
@@ -411,22 +411,22 @@ __global__ void reduce6(int *g_data_index, float *g_data_value, unsigned int set
 }
 
 
-__device__ float compute_zaehler(float dot_xi_yi, float* dot_yi_x, float* dot_xi_x, int p, int max_p_index )
+__device__ double compute_zaehler(double dot_xi_yi, double* dot_yi_x, double* dot_xi_x, int p, int max_p_index )
 {
 	//todo: samevector, kann vorberechnet werden.
-	float zaehler = dot_xi_yi - dot_yi_x[max_p_index] - dot_xi_x[max_p_index] + dot_same[p][max_p_index];//kernel(p,max_p_index, p, max_p_index);
+	double zaehler = dot_xi_yi - dot_yi_x[max_p_index] - dot_xi_x[max_p_index] + dot_same[p][max_p_index];//kernel(p,max_p_index, p, max_p_index);
 	return zaehler;
 }
 
 
-__device__ float compute_nenner(float dot_xi_xi, float* dot_xi_x, int p, int max_p_index)
+__device__ double compute_nenner(double dot_xi_xi, double* dot_xi_x, int p, int max_p_index)
 {
-	float nenner = dot_xi_xi - 2* dot_xi_x[max_p_index] +  dot_same[p][max_p_index];//kernel(p, max_p_index, p, max_p_index);
+	double nenner = dot_xi_xi - 2* dot_xi_x[max_p_index] +  dot_same[p][max_p_index];//kernel(p, max_p_index, p, max_p_index);
 	return nenner;
 }
 
 
-__device__ void add_to_weights(float* weights, float lambda, int index, int set)
+__device__ void add_to_weights(double* weights, double lambda, int index, int set)
 {
 	int i;
 	for (i=0; i<data_size[set]; i++)
@@ -439,7 +439,7 @@ __device__ void add_to_weights(float* weights, float lambda, int index, int set)
 }
 
 
-__device__ float update_xi_xi(float dot_xi_xi, float* dot_xi_x, int p, int max_p_index, float lambda)
+__device__ double update_xi_xi(double dot_xi_xi, double* dot_xi_x, int p, int max_p_index, double lambda)
 {
 	dot_xi_xi = lambda * lambda * dot_xi_xi
 		+ 2 * lambda * (1.0 - lambda) * dot_xi_x[max_p_index]
@@ -449,14 +449,14 @@ __device__ float update_xi_xi(float dot_xi_xi, float* dot_xi_x, int p, int max_p
 }
 
 
-__device__ float update_xi_yi(float dot_xi_yi, float* dot_yi_x, int max_p_index, float lambda)
+__device__ double update_xi_yi(double dot_xi_yi, double* dot_yi_x, int max_p_index, double lambda)
 {
 	dot_xi_yi = lambda * dot_xi_yi + (1.0 - lambda) * dot_yi_x[max_p_index];
 	return dot_xi_yi;
 }
 
 
-__device__ void update_xi_x(float* dot_xi_x, int p, int p2, int max_p_index, float lambda, float* computed_kernels , int tid)
+__device__ void update_xi_x(double* dot_xi_x, int p, int p2, int max_p_index, double lambda, double* computed_kernels , int tid)
 {
 	if( (tid < data_size[0] && p2 == 0) || ( tid >= data_size[0] && p2 == 1 ) )
 	{
@@ -472,7 +472,7 @@ __device__ void update_xi_x(float* dot_xi_x, int p, int p2, int max_p_index, flo
 __device__ int nr_of_cache_entries;
 __device__ int nr_of_elements;
 
-__device__ float* data;
+__device__ double* data;
 
 __device__ int* look_up_table;	 // translates data id to cache position
 								 // translates cache positions to id
@@ -486,7 +486,7 @@ __device__ int ca_free_pos;		 // safes which pos has no yet been occupied
 __device__ bool ca_cachemiss;
 
 __global__ void cuda_cache_init(int g_nr_of_cache_entries, int g_nr_of_elements,
-int *g_look_up_table, int* g_reverse_look_up_table, int* g_circular_array, float* g_data_cache)
+int *g_look_up_table, int* g_reverse_look_up_table, int* g_circular_array, double* g_data_cache)
 {
 	// cache initialisieren
 	look_up_table = g_look_up_table;
@@ -634,7 +634,7 @@ __global__ void cuda_kernel_computekernels_cache()
 				data[circular_array[ca_first] * nr_of_elements + tid] = kernel(0, max_p_index, t_set, t_element);
 
 								 //todo: cache einbauen
-			float* computed_kernels = &data[circular_array[ca_first] * nr_of_elements];
+			double* computed_kernels = &data[circular_array[ca_first] * nr_of_elements];
 
 			update_xi_x(dot_xi_x, 0, 0, max_p_index, lambda, computed_kernels, tid);
 
@@ -645,7 +645,7 @@ __global__ void cuda_kernel_computekernels_cache()
 			if( ca_cachemiss == true )
 				data[circular_array[ca_first] * nr_of_elements + tid] = kernel(1, max_q_index, t_set, t_element);
 
-			float* computed_kernels = &data[circular_array[ca_first] * nr_of_elements];
+			double* computed_kernels = &data[circular_array[ca_first] * nr_of_elements];
 
 			update_xi_x(dot_yi_y, 1, 1, max_q_index, lambda, computed_kernels, tid);
 
@@ -655,9 +655,9 @@ __global__ void cuda_kernel_computekernels_cache()
 }
 // cache ende
 
-__global__ void cuda_kernel_init_pointer(float* g_data0, float* g_data1 , int g_maximum_index, int g_data0_size, int g_data1_size, float* g_weights0, float* g_weights1 ,
-float *g_dot_xi_x, float *g_dot_yi_x, float *g_dot_xi_y, float *g_dot_yi_y,
-float* g_dot_same0, float* g_dot_same1, float* g_distance, float* g_rho, struct svm_parameter g_param)
+__global__ void cuda_kernel_init_pointer(double* g_data0, double* g_data1 , int g_maximum_index, int g_data0_size, int g_data1_size, double* g_weights0, double* g_weights1 ,
+double *g_dot_xi_x, double *g_dot_yi_x, double *g_dot_xi_y, double *g_dot_yi_y,
+double* g_dot_same0, double* g_dot_same1, double* g_distance, double* g_rho, struct svm_parameter g_param)
 {
 	dot_xi_x = g_dot_xi_x;
 	dot_yi_x = g_dot_yi_x;
@@ -770,8 +770,8 @@ __global__ void cuda_kernel_lambda()
 {
 	if (max_p >= max_q)
 	{
-		float zaehler = compute_zaehler(dot_xi_yi, dot_yi_x, dot_xi_x, 0, max_p_index);
-		float nenner = compute_nenner(dot_xi_xi, dot_xi_x, 0, max_p_index);
+		double zaehler = compute_zaehler(dot_xi_yi, dot_yi_x, dot_xi_x, 0, max_p_index);
+		double nenner = compute_nenner(dot_xi_xi, dot_xi_x, 0, max_p_index);
 
 		lambda = zaehler / nenner;
 
@@ -824,7 +824,7 @@ __global__ void cuda_kernel_distance()
 
 	//todo: rueckgabewert an host zurueckgeben
 
-	//float adg = max_p + max_q;
+	//double adg = max_p + max_q;
 
 	//printf("max_p = %f  max_q = %f ", max_p, max_q);
 	//printf("adg = %f ", adg);
@@ -833,10 +833,10 @@ __global__ void cuda_kernel_distance()
 	// adg / ||p-q||^2 - adg
 	// adg / <p-q, p-q> - adg
 
-	//float distance = dot_xi_xi + dot_yi_yi - 2 * dot_xi_yi;
+	//double distance = dot_xi_xi + dot_yi_yi - 2 * dot_xi_yi;
 
-	//float rdg_nenner = distance - adg;
-	//float rdg;
+	//double rdg_nenner = distance - adg;
+	//double rdg;
 
 	//if (rdg_nenner <= 0)
 	//{
@@ -855,7 +855,7 @@ __global__ void cuda_kernel_distance()
 	//print_weights(y_weights, prob[1]);
 
 	//rho = - dot_xi_yi + dot_xi_xi - (dot_xi_xi + dot_yi_yi - 2 * dot_xi_yi)/2;
-	//float rho = dot_xi_yi - dot_xi_xi - (dot_xi_xi + dot_yi_yi - 2 * dot_xi_yi)/2;
+	//double rho = dot_xi_yi - dot_xi_xi - (dot_xi_xi + dot_yi_yi - 2 * dot_xi_yi)/2;
 	//printf("xi_xi = %f   yi_yi = %f   xi_yi = %f \n", dot_xi_xi, dot_yi_yi, dot_xi_yi);
 
 }
